Added a GitHub action to compile the translation tables for us at every commit to master
