+++ .github/workflows/compile-translation-tables.yml
@@ -1,0 +1,29 @@
+name: Compile the translation tables
+env:
+  GITHUB_TOKEN: ${{ secrets.github_token }}
+  COMMIT_MSG: Compiled the translation tables
+    skip-checks: true
+on: 
+  push:
+    branches:
+      - master
+jobs:
+  compile:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Install Dependencies
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.9
+        run: pip3 install PyYAML
+      - name: Compile the translation tables
+        run: admin-tools/compile-translation-tables.py
+      - name: Commit to the repository
+        run: |
+          git config user.name "GitHub Actions Robot"
+          git remote set-url origin "https://x-access-token:${GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git"
+          git checkout master
+          git add .
+          git commit -m "${COMMIT_MSG}"
+          git diff --quiet && git diff --staged --quiet || (git commit -m "${COMMIT_MSG}"; git push origin master)
+
