+++ .gitignore
@@ -23,5 +23,4 @@ _Copies_/
 _Database_/
 build/
 dist/
-mathics_scanner/data/characters.json
 tmp
